+++ Ejercicio 3/real.h
@@ -7,11 +7,11 @@ using namespace std;
 
 class Real final : public Number {
     private:
-        const float value;
+        const double value;
 
     public:
         Real() : value(0) {};
-        Real(const float _value) : value(_value) {};
+        Real(const double _value) : value(_value) {};
 
         unique_ptr<Number> operator+(const Number& number) const override;
         unique_ptr<Number> operator-(const Number& number) const override;
